+++ src/control-flow/node-mapper.ts
@@ -1,5 +1,5 @@
 import type Parser from "web-tree-sitter";
-import { inplaceAddRange, newRanges, type SimpleRange } from "./ranges";
+import { type SimpleRange, inplaceAddRange, newRanges } from "./ranges";
 /**
  * This module provides the facilities for matching code offsets to CFG nodes.
  *