+++ src/test/commentTestCollector.ts
@@ -6,9 +6,9 @@
  */
 import { Glob } from "bun";
 import { getTestFuncs as getTestFuncsForC } from "./collect-c";
+import { getTestFuncs as getTestsForCpp } from "./collect-cpp.ts";
 import { getTestFuncs as getTestFuncsForGo } from "./collect-go";
 import { getTestFuncs as getTestsForPython } from "./collect-python";
-import { getTestFuncs as getTestsForCpp } from "./collect-cpp.ts";
 import type { TestFunction } from "./commentTestTypes";
 
 export const testsDir = `${import.meta.dir}/commentTestSamples`;