+++ src/components/Demo.svelte
@@ -20,6 +20,10 @@
     "C++": "void main() {\n\tif (x) {\n\t\treturn;\n\t}\n}",
     Python: "def example():\n    if x:\n        return",
     TypeScript: "function main() {\n\tif (x) {\n\t\treturn;\n\t}\n}",
+    TSX: "function getGreeting(user) {\n" +
+      "  if (user) {\n" +
+      "    return <h1>Hello, {formatName(user)}!</h1>;  }\n" +
+      "  return <h1>Hello, Stranger.</h1>;}",
   };
 
   export let code: { [language in Language]?: string } = {};
@@ -38,15 +42,16 @@
     { language: "C" as Language, text: "C", codeMirror: cpp },
     {
       language: "Python" as Language,
-      text: "Python (experimental)",
+      text: "Python",
       codeMirror: python,
     },
     {
       language: "C++" as Language,
-      text: "C++ (experimental)",
+      text: "C++",
       codeMirror: cpp,
     },
-    {language:"TypeScript" as Language, text:"TypeScript (experimental)", codeMirror: ()=>javascript({typescript:true})}
+    {language:"TypeScript" as Language, text:"TypeScript (experimental)", codeMirror: ()=>javascript({typescript:true})},
+    {language:"TSX" as Language, text:"TSX (experimental)", codeMirror: ()=>javascript({typescript:true, jsx:true})},
   ] as const;
 
   const urlParams = new URLSearchParams(window.location.search);
@@ -73,6 +78,11 @@
   if (urlParams.has("typescript")) {
     languageCode.TypeScript = LZString.decompressFromEncodedURIComponent(
       urlParams.get("typescript"),
+    );
+  }
+  if (urlParams.has("tsx")) {
+    languageCode.TSX = LZString.decompressFromEncodedURIComponent(
+      urlParams.get("tsx"),
     );
   }