+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+
+- A simple CLI script to render a single function from a file (`scripts/render-function.ts`)
+- [Developer docs](https://tmr232.github.io/function-graph-overview/docs)
+
 ## [0.0.11]
 
 ### Added