+++ CHANGELOG.md
@@ -11,6 +11,15 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - A simple CLI script to render a single function from a file (`scripts/render-function.ts`)
 - [Developer docs](https://tmr232.github.io/function-graph-overview/docs)
 - Support for click-to-navigate in the [JetBrains plugin](https://github.com/tmr232/jb-function-graph-overview).
+- Support for TypeScript
+
+### Changed
+
+- Unified a lot of the statement processing code between different languages
+
+### Fixed
+
+- `continue label` and `break label` now work properly in Go
 
 ## [0.0.11]